requirements-linux-cuda.txt
absl-py==2.0.0
accelerate==0.23.0
addict==2.4.0
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
autopep8==2.0.4
basicsr==1.4.2
billiard==4.1.0
bitsandbytes==0.41.1
cachetools==5.3.1
celery==5.3.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.1.1
cycler==0.12.1
diffusers @ git+https://github.com/huggingface/diffusers@bc8a08f67ce0faf89a0a9f8ecc0ec6b6253b5966
exceptiongroup==1.1.3
facexlib==0.3.0
fastapi==0.104.0
filelock==3.9.0
filterpy==1.4.5
fire==0.5.0
fonttools==4.43.1
fsspec==2023.9.2
future==0.18.3
gfpgan==1.3.8
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.2
lazy_loader==0.3
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.5
MarkupSafe==2.1.2
matplotlib==3.8.0
minio==7.1.17
mpmath==1.3.0
networkx==3.0
numba==0.58.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
realesrgan==0.3.0
redis==5.0.1
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tb-nightly==2.15.0a20231019
tensorboard-data-server==0.7.1
termcolor==2.3.0
tifffile==2023.9.26
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
xformers==0.0.22.post7
yapf==0.40.2
zipp==3.17.0
